+++ pkg/backup/backupCreateManager_test.go
@@ -18,7 +18,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-var testCtx = context.TODO()
+var testCtx, _ = context.WithTimeout(context.TODO(), 1)
 
 func TestNewBackupCreateManager(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
@@ -28,7 +28,7 @@ func TestNewBackupCreateManager(t *testing.T) {
 		blueprintInterface := newMockBlueprintInterface(t)
 
 		// when
-		manager := newBackupCreateManager(clientMock, clientSetMock, blueprintInterface, "", nil)
+		manager := newBackupCreateManager(clientMock, clientSetMock, blueprintInterface, "", nil, 1)
 
 		// then
 		require.NotNil(t, manager)
@@ -43,7 +43,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(nil)
 
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
@@ -105,7 +105,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
@@ -455,7 +455,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(assert.AnError)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(assert.AnError)
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
@@ -516,7 +516,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(assert.AnError)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(assert.AnError)
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
@@ -579,7 +579,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
@@ -639,7 +639,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
@@ -695,7 +695,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
@@ -755,7 +755,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
-		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
+		providerMock.EXPECT().CreateBackup(mock.Anything, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
 		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock