+++ pkg/backup/backupManager_test.go
@@ -11,10 +11,12 @@ func TestNewBackupManager(t *testing.T) {
 		// given
 		clientSetMock := newMockEcosystemInterface(t)
 		clientMock := newMockK8sClient(t)
+		backupRetryTimeLimit := 10
+
 		blueprintInterface := newMockBlueprintInterface(t)
 
 		// when
-		manager := NewBackupManager(clientMock, clientSetMock, blueprintInterface, testNamespace, nil)
+		manager := NewBackupManager(clientMock, clientSetMock, blueprintInterface, testNamespace, nil, backupRetryTimeLimit)
 
 		// then
 		require.NotNil(t, manager)